move cartographer player profile code into it's own folder
