+++ xlive/Blam/Engine/camera/first_person_camera.cpp
@@ -5,7 +5,7 @@
 #include "observer.h"
 #include "game/game_globals.h"
 #include "game/player_control.h"
-#include "saved_games/cartographer_player_profile.h"
+#include "saved_games/cartographer_player_profile/cartographer_player_profile.h"
 #include "units/units.h"
 #include "units/unit_definitions.h"
 #include "units/vehicle_definitions.h"
