+++ xlive/Blam/Engine/game/player_control.cpp
@@ -7,7 +7,7 @@
 #include "game/game_time.h"
 #include "H2MOD/Modules/CustomVariantSettings/CustomVariantSettings.h"
 #include "main/interpolator.h"
-#include "saved_games/cartographer_player_profile.h"
+#include "saved_games/cartographer_player_profile/cartographer_player_profile.h"
 #include "units/units.h"
 
 real32 g_player_control_dt = 0.0f;
