+++ xlive/Blam/Engine/input/input_xinput.cpp
@@ -5,7 +5,7 @@
 #include "game/game.h"
 #include "game/game_time.h"
 #include "networking/logic/life_cycle_manager.h"
-#include "saved_games/cartographer_player_profile.h"
+#include "saved_games/cartographer_player_profile/cartographer_player_profile.h"
 #include "H2MOD/Modules/Input/ControllerInput.h"
