+++ xlive/H2MOD/GUI/imgui_integration/AdvancedSettings.cpp
@@ -15,7 +15,7 @@
 #include "networking/messages/network_message_type_collection.h"
 #include "networking/logic/life_cycle_manager.h"
 #include "rasterizer/rasterizer_globals.h"
-#include "saved_games/cartographer_player_profile.h"
+#include "saved_games/cartographer_player_profile/cartographer_player_profile.h"
 
 #include "H2MOD.h"
 #include "H2MOD/Modules/CustomMenu/CustomLanguage.h"
