+++ xlive/Project_Cartographer.vcxproj
@@ -1155,7 +1155,7 @@
     <ClCompile Include="Blam\Engine\render\render_visible_geometry.cpp" />
     <ClCompile Include="Blam\Engine\render\render_water.cpp" />
     <ClCompile Include="Blam\Engine\render\render_weather.cpp" />
-    <ClCompile Include="Blam\Engine\saved_games\cartographer_player_profile.cpp" />
+    <ClCompile Include="Blam\Engine\saved_games\cartographer_player_profile\cartographer_player_profile.cpp" />
     <ClCompile Include="Blam\Engine\saved_games\game_state.cpp" />
     <ClCompile Include="Blam\Engine\saved_games\game_state_procs.cpp" />
     <ClCompile Include="Blam\Engine\render\render_sky.cpp" />
@@ -2271,7 +2271,7 @@
     <ClInclude Include="Blam\Engine\render\render_water.h" />
     <ClInclude Include="Blam\Engine\render\render_weather.h" />
     <ClInclude Include="Blam\Engine\render\weather_definitions.h" />
-    <ClInclude Include="Blam\Engine\saved_games\cartographer_player_profile.h" />
+    <ClInclude Include="Blam\Engine\saved_games\cartographer_player_profile\cartographer_player_profile.h" />
     <ClInclude Include="Blam\Engine\saved_games\saved_game_files_async_windows.h" />
     <ClInclude Include="Blam\Engine\saved_games\game_state.h" />
     <ClInclude Include="Blam\Engine\saved_games\game_state_procs.h" />
